@@ -1,3 +1,3 @@
-class Identity < ActiveRecord::Base
+class Identity < ApplicationRecord
   belongs_to :user
 end
